Fixed broken getResponseBytes() when using Request download()
Accidentally introduced a bug where using any of the download() methods from Request would read and close the InputStream containing the response body. This caused subsequent calls to Response getResponseBytes() or getResponseText() to return nothing. Now, bytes get written to the Response object as they are read from the InputStream.